+++ src/lib/vterm.h
@@ -84,15 +84,15 @@ public:
 	void expose(u16 x, u16 y, u16 w, u16 h);
 	void inverse(u16 sx, u16 sy, u16 ex, u16 ey);
 
-	u16 charCode(u16 x, u16 y) { return text[get_line(y) * max_width + x]; }
+	u32 charCode(u16 x, u16 y) { return text[get_line(y) * max_width + x]; }
 	CharAttr charAttr(u16 x, u16 y) { return attrs[get_line(y) * max_width + x]; }
 
 	static s32 charWidth(u32 ucs);
 
 protected:
-	virtual void drawChars(CharAttr attr, u16 x, u16 y, u16 w, u16 num, u16 *chars, bool *dws) = 0;
+	virtual void drawChars(CharAttr attr, u16 x, u16 y, u16 w, u16 num, u32 *chars, bool *dws) = 0;
 	virtual bool moveChars(u16 sx, u16 sy, u16 dx, u16 dy, u16 w, u16 h) { return false; }
-	virtual void drawCursor(CharAttr attr, u16 x, u16 y, u16 c) {}
+	virtual void drawCursor(CharAttr attr, u16 x, u16 y, u32 c) {}
 	virtual void sendBack(const s8 *data) {}
 	virtual void modeChanged(ModeType type) {}
 	virtual void historyChanged(u32 cur, u32 total) {}
@@ -225,7 +225,7 @@ private:
 	CharsetMap s_g0_charset, s_g1_charset;
 
 	// terminal info
-	u16 *text;
+	u32 *text;
 	CharAttr *attrs;
 	s8 *tab_stops;
 	u16 *linenumbers;
